fix: robust draco interface include directories and clear last warnings

+++ src/gltflib/parsemeshdata.cpp
@@ -43,7 +43,7 @@ void ParseMeshdata::getAllMeshes()
             if (iter_primitives->material >= 0 && (size_t)iter_primitives->material < model->materials.size()) {
                 primitive.material = model->materials[iter_primitives->material];
             }
-            int positionSize                      = 0;
+            size_t positionSize                      = 0;
             std::map<std::string, int> attributes = iter_primitives->attributes;
             if (attributes.find("POSITION") != attributes.end())
             {
